Extracted test_suite_sorting
